+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 * `plom-server launch foo` starts a plom server in the directory `foo`.
+* `plom-server` has new command line args to control logging.
 * new `PlomServer`, `PlomDemoServer`, and `PlomLiteDemoServer` objects for interactively running servers, or otherwise running a server in a background process.
 * `plom-hwscan` can now specify precise per-page mappings from the bundle to questions on the server.